+++ src/main/java/com/qa/user_app/data/repository/PokemonRepository.java
@@ -5,7 +5,10 @@ import org.springframework.stereotype.Repository;
 
 import com.qa.user_app.data.entity.Pokemon;
 
+// Declare interface which extends Spring Data JPA's JpaRepository
+// Domain type as Pokemon and if type as Long
+
 @Repository
-public interface PokemonRepository extends JpaRepository<Pokemon, Integer>{
+public interface PokemonRepository extends JpaRepository<Pokemon, Long>{
 
 }
